pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "city-change-detector"
version = "0.1.0"
description = "Registration + semantic change detection for multi-temporal aerial imagery"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "opencv-python>=4.8",
  "Pillow>=10.0",
  "torch>=2.2",
  "transformers>=4.40",
  "huggingface_hub>=0.23",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
